+++ src/modules/location-catch-stats/settings/index.js
@@ -1,0 +1,13 @@
+/**
+ * Add settings for the module.
+ *
+ * @return {Array} The settings for the module.
+ */
+export default async () => {
+  return [
+    {
+      id: 'location-catch-stats.show-misses',
+      title: 'Show number of misses',
+    },
+  ];
+};